Improve mobile game view, handle game for 2 gates

+++ packages/table/src/services/CameraService.js
@@ -27,15 +27,18 @@ const saveReplay = async (replayDir) => {
   const pgrep = spawnSync('pgrep', ['raspivid']);
   const pid = Number(pgrep.stdout.toString());
 
-  try {
-    execSync(`kill -USR1 ${pid}`);
-    execSync(`MP4Box -add ${fileNameToRemove}.h264 ${videoFileName}.mp4 -fps 50`);
-    execSync(`rm ${fileNameToRemove}.h264`);
-  } catch (error) {
-    console.log('Error in CameraService -> saveReplay: ', error);
-  }
+  // eslint-disable-next-line no-restricted-globals
+  if (!isNaN(pid) && pid !== 0) {
+    try {
+      execSync(`kill -USR1 ${pid}`);
+      execSync(`MP4Box -add ${fileNameToRemove}.h264 ${videoFileName}.mp4 -fps 50`);
+      execSync(`rm ${fileNameToRemove}.h264`);
+    } catch (error) {
+      console.log('Error in CameraService -> saveReplay: ', error);
+    }
 
-  return googleDriveService.uploadFile(videoFileName);
+    return googleDriveService.uploadFile(videoFileName);
+  }
 };
 
 const startVideoStream = (replayTime, replayDir) => {
